@@ -22,9 +22,8 @@
     "pako": "^2.1.0",
     "prop-types": "^15.8.1",
     "radium": "^0.26.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-key-handler": "^1.2.0-beta.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.3",
     "react-waypoint": "^10.3.0"
   },
